requirements.txt
fastapi[all]
python-dotenv
requests
Pillow
openai
